chore(mobile): organize components
