+++ apps/mobile/src/features/profile/components/stat-label.tsx
